created a new endpoint btn, added API method for new endpoint, added headers to modal text
